+++ frontend/public/js/controllers/singleNews_controller.js
@@ -7,19 +7,20 @@ StudyManager.SingleNewsController = Ember.Controller.extend({
             var that = this;
             var name = thisNews.get('title');
 
+            if (confirm('Wollen Sie die News \"' + name + '\" wirklich löschen?')) {
+                thisNews.deleteRecord();
 
-            thisNews.deleteRecord();
-
-            thisNews.save().then(function(response) {
-                that.transitionToRoute('news').then(function () {
-                    var aMessage = 'News \"' + name + '\" erfolgreich gelöscht!';
-                    that.get('controllers.news').set('statusMessage', { message: aMessage, isSuccess: true });
+                thisNews.save().then(function(response) {
+                    that.transitionToRoute('news').then(function () {
+                        var aMessage = 'News \"' + name + '\" erfolgreich gelöscht!';
+                        that.get('controllers.news').set('statusMessage', { message: aMessage, isSuccess: true });
+                    });
+                }, function(error) {
+                    thisNews.rollback();
+                    var aMessage = 'News \"' + name + '\" konnte nicht gelöscht werden!';
+                    that.set('statusMessage', { message: aMessage, isSuccess: false });
                 });
-            }, function(error) {
-                thisNews.rollback();
-                var aMessage = 'News \"' + name + '\" konnte nicht gelöscht werden!';
-                that.set('statusMessage', { message: aMessage, isSuccess: false });
-            });
+            }
         },
 
         updateNews: function(newData) {
